pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "openai>=1.99.6",
    "psycopg2-binary>=2.9.10",
    "oauthlib>=3.3.1",
    "requests>=2.32.4",
    "google-api-python-client>=2.178.0",
    "youtube-transcript-api>=1.2.2",
    "yt-dlp>=2025.8.11",
    "sqlalchemy>=2.0.42",
    "werkzeug>=3.1.3",
    "resend>=2.12.0",
    "google-auth>=2.40.3",
]
